chore(thread): address thread review feedback
